Revert "Committing integration tests/merge"
This reverts commit 3a9a41f82be976080ec8ad37c35e3b22e3e10878

+++ capstone/tenmo-client/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -5,14 +5,14 @@ import java.math.BigDecimal;
 public class Account {
 
     int accountId;
-    BigDecimal balance;
+    Double balance;
     int userId;
 
-    public BigDecimal getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(BigDecimal balance) {
+    public void setBalance(Double balance) {
         this.balance = balance;
     }
